refactor: separate router module
