+++ src/router/admin.go
@@ -1,27 +1,13 @@
-package server
+package router
 
 import (
 	"github.com/jinzhu/gorm"
 	"github.com/labstack/echo"
 
-	"github.com/KotaTanaka/echo-api-sandbox/handler"
 	adminhandler "github.com/KotaTanaka/echo-api-sandbox/handler/admin"
-	clienthandler "github.com/KotaTanaka/echo-api-sandbox/handler/client"
 )
 
-func Router(e *echo.Echo, db *gorm.DB) {
-	// API仕様書の出力
-	e.File("/doc", "app/redoc.html")
-	// Hello, World!
-	e.GET("/", handler.Hello())
-	// CA-01 エリアマスタ取得
-	e.GET("/areas", clienthandler.GetAreaMasterClient(db))
-	// CS-01 エリアに紐づく店舗一覧取得
-	e.GET("/shops", clienthandler.GetShopListClient(db))
-	// CR-01 店舗に紐づくレビュー一覧取得
-	e.GET("/reviews", clienthandler.GetReviewListClient(db))
-	// CR-02 店舗へのレビュー投稿
-	e.POST("/reviews", clienthandler.CreateReviewClient(db))
+func AdminRouter(e *echo.Echo, db *gorm.DB) {
 	// AA-01 エリア登録
 	e.POST("/admin/areas", adminhandler.RegisterAreaAdmin(db))
 	// AA-02 エリア削除
